Generic collision detection test suite
